LabelIcon->LegendThresholdVertical
